Updated openstack/openstack Project: openstack/glance  38dd10bb98a480d6fff9f1cf507e5b25e5ac10e6
Glance scrubber should page thru images from registry

If the registry is limited in it's page size when returning the
list of images, the scrubber will only receive the first page.
This patch enables the scrubber to retrieve all the pages of
images.

The new tests are implemented with mock rather than mox. There
is another patch (https://review.openstack.org/#/c/111995) to
convert the existing code in test_scrubber to use mock.

Change-Id: I106f679be96bd39167337c4e27f4046f90d7d7ff
Closes-Bug: 1280343