Updated openstack/openstack Project: openstack/trove  be5e7454f0d698d93678ca854d975e6f36256a91
Allow users the ability to update an instance name

Currently the users do not have the ability to rename an
instance. The current PUT call, does not allow the user
to modify individual attributes of the instance.
Hence this review introduces a new PATCH call for updating
instance name and/or configurations.
The long term plan is to deprecate PUT in favor of PATCH.

Implements: blueprint Allow users the ability to update the instance name
Author: Iccha Sethi <iccha.sethi@rackspace.com>
Co-Authored-By: Theron Voran <theron.voran@rackspace.com>
Co-Authored-By: Riddhi Shah <ridhi.j.shah@gmail.com>

Change-Id: I25c3b0906191afdc587f5ddd5777b3185dc080d4