Updated openstack/openstack Project: openstack/heat  af9b66ed513da815032f297b601321d587ee7a87
Put a cap on our cyclomatic complexity

flake8 has support for cyclomatic complexity (Mccabe) currently our
worst offender has a complexity of 28 (25 is considered very bad). So
set our max-complexity to 29 so we at least don't make things any worse.

This is the first step in bringing this number down to 20 or so.

https://github.com/flintwork/mccabe

Change-Id: Ic90093ab80d9d3e27c257ea1a5cc393e09ebc7ee