Updated openstack/openstack Project: openstack/nova  d140b42e9071b68ce835b6139776c8563e092022
Minor refactor of _setup_instance_group()

This patch is a minor refactoring of the _setup_instance_group()
method in the filter scheduler.  The majority of the method was inside
of a conditional.  This primarily reverses that and reduces the
nesting depth used.  There should be no functional changes at all in
this patch.

Change-Id: Iab4d63200653a4320c91dc1356091ff9d961cafd
Related-bug: #1379451