Updated openstack/openstack Project: openstack/keystone  1e6a9b5d4c936c1d6c55c15bb3385e819c3653af
Add max-complexity to pep8 for Keystone

The current highest complexity is 23, set the maximum complexity
for flake8 to 24. This should help to keep code simple and
maintainable. It should be possible to reduce the maximum
complexity to something closer to 20 with relatively small efforts.

keystone.common.sql.migrate_repo.versions.039_grant_to_assignment:
    'downgrade_assignment_table' is complexity of 21

keystone.token.controllers:
    'Auth._authenticate_token' is complexity of 23

Change-Id: I47cb13bac97e9144ef004977942b9eaa50a17b39