Updated openstack/openstack Project: openstack/nova  9fd059b938a2acca8bf5d58989c78d834fbb0ad8
Put a cap on our cyclomatic complexity

flake8 has support for cyclomatic complexity (Mccabe) currently our
worst offender has a complexity of 67 (25 is considered very bad). So
set our max-complexity to 68 so we at least don't make things any worse.

This is the first step in bringing this number down to 25 or so.

https://github.com/flintwork/mccabe

Change-Id: I5117ca58ad28888e24c16251aa5e92084ead97ca