Updated openstack/openstack Project: openstack/swift  87a82a35f40843993d16283e95043929636a2e4a
Fix bug with expirer and unicode

Looks like I wasn't careful enough last time i fixed this bug- this
is basically the same deal, just further down the line. What happens
without this bug fix is you upload a object with a unicode name and
a x-delete-after header and when it does expire the customer object
gets removed but the expirer marker object doesn't.

Change-Id: I82359c87a0919d71693e49ecf08e0f1eedc9d18e