Updated openstack/openstack Project: openstack-dev/devstack  572a4c4c3c5553beea01f519dbdcadfb9c6f78b7
Allow use of ceilometer agent coordination

Juno brings tooz-based coordination of ceilometer central, compute
and alarm agents. If CEILOMETER_COORDINATION_URL is set it will be
used as the value for a tooz coordination backend. If memcached is
chosen, its package will be installed. Other backends, such as
zookeeper can be configured, but installation is left as an exercise
for the devstacker.

In the default devstack setup having coordination will do little as
there are only one of each agent, but this makes it a bit easier for
multi-node or multi-agent setups.

Change-Id: Ib85ccd435de3bc0ae56b5fe8c2fce6c2af9ff8d0