Updated openstack/openstack Project: openstack-dev/devstack  167dd5fd691b877983a03c42adebd0fddb239b15
Move generation of baremetal VM names to lib/ironic

The create-nodes script currently receives a total # of VMs to create
and creates them all, generating their names on the fly.  This moves
that name generation to lib/ironic and makes the script create only
single VMs as directed.  This centralizes the naming of things to
lib/ironic and will make it easier to reference these things elsewhere.

Change-Id: I98e61f7188e027e690303e32aff7cd2347f6d2c2