Updated openstack/openstack Project: openstack/heat  020eddb9bfccd075a3f55713dc38b6efaa5395bd
Make HARestarter do a proper lookup of the RefID

HARestarter is a primitive prototype that is barely maintained.  Its
implementation still dates to the era when the resource name, physical
resource ID and RefId were regularly used interchangably.

This change substitutes Stack.resource_by_refid() for the previous
(incorrect) home-grown hack, so that at least the giant red flags will be
easy to grep for.

Change-Id: I9d81dfcd5d34282c8bd83bbd98350e75c1ec0eb2
Closes-Bug: #1381136