Updated openstack/openstack Project: openstack/tempest  66d8831d173cd4713bff8875bd516ad132db9070
Bump version string to coincide with kilo release

The tempest-2 tag has been pushed to mark the start of support for the
kilo release. This commit bumps the version to 3 to show the current
working tag number. Tempest-3 when it is tagged after the kilo release
will mark support for icehouse, juno, kilo, and the 'L' release based
on the current stable maintenance schedule.

Change-Id: If401356d441ef3dc82a5b8069168c4dd71c8b3d8