Updated openstack/openstack Project: openstack/keystonemiddleware  ab4e4b5189ef1143d189d7ca2543e368c982e927
Add versions to requests

Make the auth plugin work more like a regular plugin such that the
values returned from get_endpoint are version dependant and the requests
made to the session all have a version number.

For compatibility there is still no lookup of the correct versioned URL
it just appends the default.

Change-Id: I2a8258ebd584aa35a94369fe4673dfd5ee07715b