Updated openstack/openstack Project: openstack/keystonemiddleware  2a52517e68ba00a3ce8e0a0dd8c986a917135947
Use an adapter in IdentityServer

Now that we are using plugins we can use the adapter to set the default
values on requests rather than push it through _http_request.

Remove the session from the base class. The only reference to it we need
will be in the adapter and maintained by the IdentityServer.

Removed one test assertion. It only seems to test that a config value is
passed through the class and the final location for this is getting
nested deeper and deeper. If the identity_uri config value wasn't being
passed we would have a lot more test failures so I believe it's safe to
remove.

Change-Id: Id092e8f42f843dbfbc1c30589b50ba341ccf4aae