Updated openstack/openstack Project: openstack/neutron  280b9eacb7be2db2d9d2fd5cafad2814f07ac70c
Reorder operations in (l3_dvr) update floating ip

This review overrides update_floatingip (L3_NAT_dbonly_mixin)
in l3_dvr_db (L3_NAT_with_dvr_db_mixin) to reorder the garbage
collection to after the floating ip is updated and rpc called.
This was previously being called in the (already) overridden
_update_fip_assoc.

Since this call is moved, the _update_fip_assoc for l3_dvr_db
is exactly the same as l3_db and is thus removed completely.

This tidy up was created whilst looking at bug 1381617. The
intention was to mitigate the timing issues exposed by [1]
and discussed in the bug report. It seems the problem persists
with more discussion around 'properly fixing' this at [2].

[1] https://review.openstack.org/#/c/102700/
[2] https://review.openstack.org/#/c/130414/4

Change-Id: I24cef57bf394ceeb931ff48fa67f18218c734b00