Updated openstack/openstack Project: openstack/trove  101d454b70c0a386f39a1c1106ba66d86765cb7e
Update contributing.rst to include guidelines on Code Review

An important part of contributing to the Trove project is to review
changes proposed by others. To reflect the value of this activity, and
provide some guidelines on how the Trove project would like reviews to
be conducted, I have updated the contributing.rst file with a summary
of some conversations at the mid-cycle meetup.

The intent is that reviews of this proposed change will be a
discussion not only of the summary that I have attempted to capture,
but also of the change(s) being proposed.

I've opted to provide only rough guidelines and not provide exhaustive
examples of "good" and "bad" reviews (the format used by, for example,
the commit message guidelines) because the intent is not to create
rigid and cumbersome process but rather to provide guidelines that
reflect the desired purpose of the review process.

It is acceptable to -1 this change if you find typographical errors,
or have suggestions on the formatting of the document ;)

Change-Id: I09432c335282c17974cb62b6f77cd00ebcdda21f