Updated openstack/openstack Project: openstack/nova  a8612e7d93a0532a16e981a9e9c1a77c4df02e0b
Translate 'powervm' hypervisor_type to 'phyp' for scheduling

This adds a temporary translation of the 'powervm' hypervisor type
image metadata to 'phyp' for scheduling.

The powervc-driver in stackforge uses the hypervisor type of 'powervm'
for scheduling so this is needed to make that work until the driver
can get a translation shim in place to convert powervm to phyp in
image metadata.

Closes-Bug: #1370613

Change-Id: I4b20ffcb9911806db821f2902e09248dfa89d845