Updated openstack/openstack Project: openstack/ceilometer  1b1251d4865c934c1b3bda93e2e6dfc9926b5be3
Allow collector service database connection retry

At the startup collector service loads the database dispatcher,
if the DB is offline momentarily the collector service starts but
continues to run with out having the ability to re-try the DB
connection. This fix will allow the connection to be retried next
time when collector service attempts to write to the database.

Change-Id: I10fef61a9c525e1c84d8cadbca86d6bde84927ba
Closes-Bug: 1301777