Updated openstack/openstack Project: openstack/nova  15cb7c967c70aad8900d59d2f02776e2ba0297d1
Remove use of unicode on exceptions

Casting exceptions to unicode using unicode() can interfere
with the proper translation of the exception message.  This
is especially true when lazy translation is enabled, since it
causes the exception message to be translated immediately using
the default locale.

This is the same problem caused by using str on exceptions,
which was fixed by https://review.openstack.org/#/c/116054/

In addition to fixing these cases, this patch updates
hacking check N325, which checks for the use of str() on
exceptions, to also check for the use of unicode().

In order to make it clear which cases that cannot be caught
by the hacking check have been inspected, they have been
converted to using six.text_type() instead of unicode().

Closes-bug: #1380806
Change-Id: I87bb94fa76458e028beba28d092320057b53f70a