Updated openstack/openstack Project: openstack/neutron  2eef0dc03d4fd6d7363cdb8fc02c4d5d188c1a88
Move disabling of metadata and ipv6_ra to _destroy_router_namespace

I noticed that disable_ipv6_ra is called from the wrong place and that
in some cases it was called with a bogus router_id because the code
made an incorrect assumption about the context.  In other case, it was
never called because _destroy_router_namespace was being called
directly.  This patch moves the disabling of metadata and ipv6_ra in
to _destroy_router_namespace to ensure they get called correctly and
avoid duplication.

Change-Id: Ia76a5ff4200df072b60481f2ee49286b78ece6c4
Closes-Bug: #1383495