Updated openstack/openstack Project: openstack/nova  54b0da297471529e5ba575c70371fe61c9497c0b
Make pci_requests a proper field on Instance object

When we added the pci_requests object, we didn't plumb it all the way into the
Instance object because of time constraints. This finishes that work so that it
behaves properly and doesn't require an additional DB lookup every time we need to
pull it based on an instance.

Partial-Bug: #1368260
Change-Id: I134441c8288d5dbdc0bac6163cb2affc9f0af762