Updated openstack/openstack Project: openstack/nova  c5864c8c5b6d5a92f92eb0da4bf0b16dd6f1008c
Nova should be in charge of its log defaults

Oslo's log module has a set_defaults which Nova can use
to set the context logging format and default log levels.
We should not have to fix default levels in oslo and then
sync across to Nova. This becomes all the more important
when we move to oslo.log library so we should get started
now

Change-Id: Id47099d5a755bc2923b05fe4bef270a7c0f99389