Updated openstack/openstack Project: openstack-infra/devstack-gate  a34d377fb29f45559b3f63696039ecf7f3c3c741
Add is_debian and uses_debs functions

Create a new is_debian method to check for Debian
distro. Create an uses_debs function and use
it in the same cases as Ubuntu to support Debian
distros in tests. Also keeping the is_ubuntu and
is_debian methods for some special cases if needed.

Change-Id: Id30cdb9bc94591a957d6f77be84e077a0adad7d9