Updated openstack/openstack Project: openstack/nova  49acbb29804a4ddb64cde15ccded8392b4adf0ad
Reduce the complexity of _get_guest_config()

Running flake8 to check for cyclomatic complexity showed that the
libvirt driver's _get_guest_config() method was the worst offender,
with a complexity of 67. This commit simply refactors this method,
splitting off nested cases into their own methods, and reduces the
complextity to 40. No functional changes are included.

Change-Id: Ic27e091f922118edb48b233fb31c702d9528bb75