Updated openstack/openstack Project: openstack/taskflow  5d2e406a680c7734adeae33b7946cad4e38690b3
Use constants for link metadata keys

Instead of using strings (which can be easy to
mistype and get wrong), provide a set of constants
that can be used to attach and use these keys in
flows and at compilation.

This also helps make it more clear what the keys
do and where they are used.

Change-Id: I5283b27617961136a4582bbcfce4617f05e8dd1d