Updated openstack/openstack Project: openstack/python-keystoneclient  c72fa61c4d144328c8956400b934a55d746cf6dd
Remove warning about management token

Because keystoneclient tries to figure out a management_url in all
contexts, it means that any end-user who uses any python-*client from
OpenStack that is using python-keystoneclient will always get a warning
about not being able to get a management token. This is, however, not
something that they need warning about, since there is no expectation
they'll have one. It's distressing to see it as part of normal
operation.

So just remove the warning.

Change-Id: Ia103a53c09c00fc09cef5fb24be546fc1da0684a