Updated openstack/openstack Project: openstack/cinder  c2ec1beb48e7e96d50d94fe151f0387caba59c4c
Reserve 5 migrations for backports

Reserve 5 migrations incase the need arises to backport any
fixes that require a db migration in stable/juno.

We've never set this up in the past and we did run into a case
last cycle where we had to hack some things around to make it work
without the place holder.

Why 5?  Why not?  For as little as we touch the DB historically
this number should be more than sufficient.

Change-Id: I2830f6557c65c4c9b3056bb5ee72b17346e21823