Updated openstack/openstack Project: openstack/cinder  bc17a71cdbebfef597bfe950e29c8e94705bdd5c
CiscoFCSanLookupSerive uses extra argument in init

This patch fixed two issues with the __init__ routine in
CiscoFCSanLookupService:

1. There's an extra argument in super(CiscoFCSanLookupService,
self).__init__(self, **kwargs). It should be changed to
super(CiscoFCSanLookupService, self).__init__(**kwargs).

2. The last line 'self.fabric_configs = ""' should be removed.
self.fabric_configs was created in self.create_configuration() in the
middle of the __init__ routine. It shouldn't be cleared out at the end
of the __init__ routine.

Change-Id: I2623f3b51af34d14849015e52e3676271c9fe414
Closes-Bug: #1385974