Updated openstack/openstack Project: openstack/trove  271cdc0a9306b7cf94482b2c0d53cd299c0cfa65
Couchbase Root Password Can Go Out Of Sync

if a user goes to the web console of a couchbase instance and adds
another couchbase instance as a server to create a cluster, the root
password for the second server is replaced with the first server's
password (aka they're synced).

this is problematic because the ping check today relies on the
password in /etc/couchbase/secret_key, which is only set on
prepare(). this means that the status for the second server (and
the third, and etc.) flips to 'shutdown' because the state of the
database cannot be ascertained.

the solution is to retrieve the current password from config.dat if
the status check fails with the stored value in secret_key. if the
password parsed from config.dat works, update the stored value in
secret_key.

Change-Id: Ia39b9c57df88b6784d13d8f3fd86967c7cffa743
Closes-Bug: #1380984