Updated openstack/openstack Project: openstack/nova  9736fc659c8a681d86bb7683262d3f757202e5db
Use wsgi.response for v2.1 unrescue API

According to the comment on v2.1 unrescue API, v2 unrescue API seems
to return a HTTP200 response on success case. However that is invalid
and it returns a HTTP202 response like the following:

 $ curl -i 'http://nova-api/v2/{tenant-id}/servers/{server-id}/action'
   -X POST -H "Accept: application/json" [..] -d '{"unrescue": null}'
 [..]
 RESP: [202] CaseInsensitiveDict({'date': 'Tue, 21 Oct 2014 02:23:42 GMT',
   'content-length': '0', 'content-type': 'text/html; charset=UTF-8',
   'x-compute-request-id': 'req-642b49e8-0e13-4923-a49b-6c8b317b7b8c'})
 RESP BODY: null

This patch removes the comment and makes the API use wsgi.response for
the code cleanup.

Partially implements blueprint v2-on-v3-api

Change-Id: Ie78e56fafd71f58568a0562b9a452f14d3b19815