Updated openstack/openstack Project: openstack/oslo.concurrency  78ba1438794ce19c9a67d4d775e3bd113b447582
Don't use variables that aren't initialized

The code should not try to use acquire_time if it has not been
initialized to some reasonable value. This commit sets it to
none and checks that it has been set before using it in the
release method to calculate the holding time.

Change-Id: Id55d549f75a348d744f753a68e1ec298bdd25b33