Updated openstack/openstack Project: openstack/neutron  5589c978fbafd79d8f15ad48fd659fedd6362d4d
Fix 'Length too long' error in neutron-dsvm-functional tests

testr is occasionally failing with 'Length too long' for
neutron-dsvm-functional test cases.

The issue is seen when the size of the testr log file size goes into MBs.

(Please see the bug comments for the details on the testr logs)

This patch fixes the issue by disabling debug logs in the tests.

Change-Id: I48fc9d14f6886f86761038ee4d106fbbe59b6890
Closes-bug: #1382579