Updated openstack/openstack Project: openstack/requirements  508b4d2db050c193cb8478380e9b378749d1eaad
implement -s / --soft-update flag

This provides a new flag to update.py which allows for a "soft
update", which updates all the requirements found in g-r to the g-r
specified versions, but lets any unknown lines pass through as is,
without warning. This makes the tool usable for ecosystem projects
that want to test with g-r versions of the g-r specified dependencies,
but also want to be able to add their own dependencies above and beyond.

Change-Id: I1f195ef9ff1509659848e14ec9936ff6f66a6496